Add Insert, Get tests. Update Insert, Append, Get
Add unit tests for methods Insert and Get to test out if elements are
being inserted in correct order. Simplify method of CircularList class
Get by changing loop 'while' to 'for' and method Append by updating
existing loop so that the loops are only used to get to needed element
and every other operation is done outside of them. Update Insert method
with now adding the case of inserting to the 0 position, if the length
is 1 or more and changing 'while' to 'for' loop in a previously
described case for the same reason as for the other two methods